% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stigmaR-package.R
\docType{package}
\name{stigmaR-package}
\alias{stigmaR-package}
\title{stigmaR: Measuring and calculating structural stigma}
\description{
Provides researchers with pre-computed, state-level structural stigma scores derived from publicly available data. Users can merge stigma indices—including implicit bias, explicit attitudes, policies, and belief items—into their own data frames using a single function call.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/stigmaRverse/stigmaR}
  \item \url{https://stigmaRverse.github.io/stigmaR/}
  \item Report bugs at \url{https://github.com/stigmaRverse/stigmaR/issues}
}

}
\author{
\strong{Maintainer}: Seungju Kim \email{follhim@gmail.com}

Authors:
\itemize{
  \item Seungju Kim \email{follhim@gmail.com}
}

}
\keyword{internal}
