% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stigmaR-package.R
\docType{package}
\name{stigmaR-package}
\alias{stigmaR-package}
\title{stigmaR: Measuring and calculating structural stigma}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/follhim/stigmaR}
  \item \url{https://follhim.github.io/stigmaR/}
  \item Report bugs at \url{https://github.com/follhim/stigmaR/issues}
}

}
\author{
\strong{Maintainer}: Seungju Kim \email{follhim@gmail.com}

}
\keyword{internal}
